test: split the auto_cycle and auto_zoom tests out of card.test.ts
card.test.ts was 99.5K against 1-10K for every sibling. Three behaviour
groupings move out:

  card.auto-cycle.test.ts     refresh timer, visibilitychange, date follows now
  card.auto-zoom.test.ts      periodic zoom cycle and everything suspending it
  card.auto-behaviour.test.ts the contract between the two

The last one exists because every case in it asserts both behaviours at
once; filing the table under either name would make the other look
uncovered.

Naming is deliberate: the sibling files mirror source filenames
(card-config.test.ts), and there is no auto-zoom.ts, so the dotted form
marks these as behavioural groupings of card.ts rather than a missing
module.

Shared fixtures move to test/card/helpers.ts, which every card test file
now imports instead of closing over its own copies. setupCardTest() is
one call rather than two because no file ever wanted only half of it.

No source change, no test added or removed: 801 before, 801 after.
