+++ test/card/card.test.ts
@@ -1,52 +1,19 @@
-import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, describe, expect, it, vi } from "vitest";
 import { SolarViewCard } from "../../src/card/card.js";
-import { formatDate } from "../../src/card/card-template.js";
 import { EPIC_BASE_URL } from "../../src/card/gallery/source-resolver-dscovr-earth.js";
 import { getSunImageUrl } from "../../src/card/gallery/source-resolver-sdo-sun.js";
-import { UrlCache, urlCache } from "../../src/card/gallery/url-cache.js";
+import { UrlCache } from "../../src/card/gallery/url-cache.js";
+import {
+  clickButton,
+  createAndMount,
+  dragView,
+  getSvgViewBox,
+  parseViewBox,
+  setupCardTest,
+  stubImagePreload,
+} from "./helpers.js";
 
-beforeAll(() => {
-  if (!customElements.get("ha-planetary-solar-system-card-test")) {
-    customElements.define("ha-planetary-solar-system-card-test", SolarViewCard);
-  }
-});
-
-// Any mounted card with gallery.mode != "none" now fetches in the background from
-// connectedCallback, even in describes that never stub fetch. Left unstubbed, that's a real
-// network call — works in CI (real internet) but not locally, and a slow one can resolve
-// after its own test ends and pollute url-cache.ts's module-level cache for a later
-// test. Default fetch to a safe rejection for every test; individual tests override it with
-// their own vi.stubGlobal("fetch", ...) when they want specific behavior.
-//
-// Every image path (initial open, gallery thumbnail, background refresh) now preloads a
-// candidate off-DOM (ImageResolver.resolve()) before it's ever assigned to a visible <img>, via
-// `new Image()`. Left unstubbed, that never resolves in jsdom (no real network), hanging
-// every await. Default it to succeed on the next microtask for every test; pass one boolean
-// per attempt to control retries (e.g. stubImagePreload(false, true) — first attempt fails,
-// the sun retry succeeds); the last value repeats for any further attempt.
-function stubImagePreload(...results) {
-  let calls = 0;
-  vi.stubGlobal(
-    "Image",
-    class {
-      src = "";
-      decode() {
-        const succeeds = results.length ? results[Math.min(calls, results.length - 1)] : true;
-        calls++;
-        return succeeds ? Promise.resolve() : Promise.reject(new Error("decode failed"));
-      }
-    }
-  );
-}
-
-beforeEach(() => {
-  vi.stubGlobal("fetch", vi.fn().mockRejectedValue(new Error("fetch not stubbed for this test")));
-  stubImagePreload();
-});
-afterEach(() => {
-  vi.unstubAllGlobals();
-  urlCache.clear();
-});
+setupCardTest();
 
 describe("SolarViewCard", () => {
   it("should be a class", () => {
@@ -231,41 +198,6 @@ describe("SolarViewCard", () => {
       gallery: { mode: "both" },
     });
   });
-
-  function createAndMount(config) {
-    const card = document.createElement("ha-planetary-solar-system-card-test");
-    if (config) card.setConfig(config);
-    document.body.appendChild(card);
-    return card;
-  }
-
-  function clickButton(card, action) {
-    const btn = card.shadowRoot.querySelector(`button[data-action="${action}"]`);
-    btn.click();
-  }
-
-  // Drives a real pointer drag across the SVG. A zero delta taps without moving — pointerdown
-  // fires on every tap (planet clicks included), which must not count as taking the view over.
-  function dragView(card, dx = 0, dy = 0) {
-    const svg = card.shadowRoot.querySelector("#solar-view svg");
-    svg.setPointerCapture = () => {};
-    svg.releasePointerCapture = () => {};
-    svg.dispatchEvent(new PointerEvent("pointerdown", { clientX: 200, clientY: 200 }));
-    if (dx || dy) {
-      svg.dispatchEvent(new PointerEvent("pointermove", { clientX: 200 + dx, clientY: 200 + dy }));
-    }
-    svg.dispatchEvent(new PointerEvent("pointerup", { clientX: 200 + dx, clientY: 200 + dy }));
-  }
-
-  function getSvgViewBox(card) {
-    const svg = card.shadowRoot.querySelector("#solar-view svg");
-    return svg.getAttribute("viewBox");
-  }
-
-  function parseViewBox(card) {
-    const parts = getSvgViewBox(card).split(" ").map(Number);
-    return { minX: parts[0], minY: parts[1], width: parts[2], height: parts[3] };
-  }
 
   describe("auto-fit viewBox", () => {
     it("sets a viewBox on the SVG on first render", () => {
@@ -951,162 +883,6 @@ describe("SolarViewCard", () => {
     });
   });
 
-  describe("refresh_mins configuration", () => {
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    it("defaults to 60000ms when refresh_mins is not set", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({});
-      expect(card._refreshMs).toBe(60000);
-    });
-
-    it("uses configured refresh_mins value", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ refresh_mins: 5 });
-      expect(card._refreshMs).toBe(300000);
-    });
-
-    it("clamps refresh_mins below 0.1 to default", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ refresh_mins: 0.05 });
-      expect(card._refreshMs).toBe(60000);
-    });
-
-    it("ignores non-numeric refresh_mins", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ refresh_mins: "abc" });
-      expect(card._refreshMs).toBe(60000);
-    });
-
-    it("timer uses configured interval", () => {
-      vi.useFakeTimers({ now: new Date("2026-02-15T10:00:00") });
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ refresh_mins: 2 });
-      card._dateNav.currentDate = new Date("2026-02-15T10:00:00");
-      document.body.appendChild(card);
-      const dateBefore = formatDate(card._dateNav.currentDate);
-      // At 60s nothing should have changed yet (interval is 120s)
-      vi.advanceTimersByTime(60000);
-      expect(formatDate(card._dateNav.currentDate)).toBe(dateBefore);
-      // At 120s the timer should fire
-      vi.advanceTimersByTime(60000);
-      expect(formatDate(card._dateNav.currentDate)).toContain("26-02-15");
-      card.remove();
-    });
-
-    it("recreates timer on setConfig when already connected", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ refresh_mins: 1 });
-      document.body.appendChild(card);
-      const firstTimer = card._autoUpdateTimer;
-      card.setConfig({ refresh_mins: 2 });
-      expect(card._autoUpdateTimer).not.toBe(firstTimer);
-      card.remove();
-    });
-  });
-
-  describe("periodic_zoom_change configuration", () => {
-    it("defaults to false when not set", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({});
-      expect(card._zoom.periodicZoomChange).toBe(false);
-    });
-
-    it("is true when configured as true", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      expect(card._zoom.periodicZoomChange).toBe(true);
-    });
-  });
-
-  describe("periodic zoom cycling", () => {
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    it("does not change zoom when periodic_zoom_change is false", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: false });
-      document.body.appendChild(card);
-      expect(card._zoomLevel).toBe(1);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(1);
-      card.remove();
-    });
-
-    it("advances zoom by one level per tick", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      expect(card._zoomLevel).toBe(1);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(3);
-      card.remove();
-    });
-
-    it("wraps from MAX_ZOOM back to level 1", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true, default_zoom: 4 });
-      document.body.appendChild(card);
-      expect(card._zoomLevel).toBe(4);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(1);
-      card.remove();
-    });
-
-    it("updates zoom level display on auto-cycle", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      expect(card.shadowRoot.querySelector(".zoom-level").textContent).toBe("1");
-      vi.advanceTimersByTime(60000);
-      // Re-query after timer fires because _render() may rebuild the DOM
-      expect(card.shadowRoot.querySelector(".zoom-level").textContent).toBe("2");
-      card.remove();
-    });
-
-    it("a refresh tick leaves a hand-set zoom level alone", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      clickButton(card, "zoom-in");
-      expect(card._zoomLevel).toBe(2);
-
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      card.remove();
-    });
-
-    it("Now button resumes the auto-cycle from the default zoom", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      clickButton(card, "zoom-in");
-      clickButton(card, "zoom-in");
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(3); // suspended by the manual zooms
-
-      clickButton(card, "today");
-      expect(card._zoomLevel).toBe(1);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      card.remove();
-    });
-  });
-
   describe("Now button highlights when the view is off-default", () => {
     afterEach(() => {
       vi.useRealTimers();
@@ -1161,256 +937,6 @@ describe("SolarViewCard", () => {
 
       vi.advanceTimersByTime(60000);
       expect(nowBtn(card).classList.contains("active")).toBe(true);
-      card.remove();
-    });
-  });
-
-  // Locks in the full "which user action stops which automatic behaviour" contract:
-  //   auto_cycle = the date following now on each refresh tick (DateNavigation._isLiveMode)
-  //   auto_zoom  = the periodic zoom cycle on each refresh tick (periodic_zoom_change)
-  // The gallery's refresh and slide timers are a separate concern and are covered elsewhere.
-  describe("what each user action stops", () => {
-    const START = new Date("2026-02-15T12:00:00Z");
-    const TICK_MS = 60000;
-
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    function mountCycling() {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      return card;
-    }
-
-    // Both behaviours are observed through what one refresh tick does, never through a flag:
-    // auto_cycle is alive if the tick pulled the date to now, auto_zoom if it moved the level.
-    function runOneTick(card) {
-      const zoomBefore = card._zoomLevel;
-      vi.advanceTimersByTime(TICK_MS);
-      return {
-        autoCycleAlive: card._dateNav.currentDate.getTime() === Date.now(),
-        autoZoomAlive: card._zoomLevel !== zoomBefore,
-      };
-    }
-
-    it("a fresh card has both running", () => {
-      vi.useFakeTimers({ now: START });
-      const card = mountCycling();
-      expect(runOneTick(card)).toEqual({ autoCycleAlive: true, autoZoomAlive: true });
-      card.remove();
-    });
-
-    // zoom-out at level 1 can't move the view, but it's still a deliberate press: intent
-    // suspends the cycle, not movement.
-    const MATRIX = [
-      { action: "zoom-in", stopsAutoCycle: false, stopsAutoZoom: true },
-      { action: "zoom-out", stopsAutoCycle: false, stopsAutoZoom: true },
-      { action: "hour-back", stopsAutoCycle: true, stopsAutoZoom: true },
-      { action: "hour-forward", stopsAutoCycle: true, stopsAutoZoom: true },
-      { action: "day-back", stopsAutoCycle: true, stopsAutoZoom: true },
-      { action: "day-forward", stopsAutoCycle: true, stopsAutoZoom: true },
-      { action: "month-back", stopsAutoCycle: true, stopsAutoZoom: true },
-      { action: "month-forward", stopsAutoCycle: true, stopsAutoZoom: true },
-    ];
-
-    for (const { action, stopsAutoCycle, stopsAutoZoom } of MATRIX) {
-      it(`${action} stops auto_cycle: ${stopsAutoCycle}, auto_zoom: ${stopsAutoZoom}`, () => {
-        vi.useFakeTimers({ now: START });
-        const card = mountCycling();
-        clickButton(card, action);
-        expect(runOneTick(card)).toEqual({
-          autoCycleAlive: !stopsAutoCycle,
-          autoZoomAlive: !stopsAutoZoom,
-        });
-        card.remove();
-      });
-    }
-
-    it("dragging the view stops auto_zoom only", () => {
-      vi.useFakeTimers({ now: START });
-      const card = mountCycling();
-      dragView(card, 50, -20);
-
-      expect(runOneTick(card)).toEqual({ autoCycleAlive: true, autoZoomAlive: false });
-      card.remove();
-    });
-
-    it("a tap that never moves the view stops neither", () => {
-      vi.useFakeTimers({ now: START });
-      const card = mountCycling();
-      dragView(card);
-
-      expect(runOneTick(card)).toEqual({ autoCycleAlive: true, autoZoomAlive: true });
-      card.remove();
-    });
-
-    it("Now restarts both after an action that stopped both", () => {
-      vi.useFakeTimers({ now: START });
-      const card = mountCycling();
-      clickButton(card, "day-forward");
-      expect(runOneTick(card)).toEqual({ autoCycleAlive: false, autoZoomAlive: false });
-
-      clickButton(card, "today");
-      expect(runOneTick(card)).toEqual({ autoCycleAlive: true, autoZoomAlive: true });
-      card.remove();
-    });
-
-    describe("replay", () => {
-      // Replay runs 36 steps inside 5s, so with the default 60s refresh a tick lands mid-replay
-      // roughly 8% of the time. Start it 3s before the tick to hit that window deterministically.
-      function startReplayJustBeforeATick(card) {
-        vi.advanceTimersByTime(57000);
-        clickButton(card, "replay");
-        vi.advanceTimersByTime(3100);
-      }
-
-      it("a tick during a replay moves neither the date nor the zoom", () => {
-        vi.useFakeTimers({ now: START });
-        const card = mountCycling();
-        startReplayJustBeforeATick(card);
-
-        expect(card._dateNav.isReplaying).toBe(true);
-        expect(card._zoomLevel).toBe(1); // the cycle must not zoom mid-animation
-        // The date belongs to the replay while it runs, not to the refresh tick.
-        expect(card._dateNav.currentDate.getTime()).toBeLessThan(START.getTime());
-        card.remove();
-      });
-
-      it("both resume on the first tick after the replay finishes", () => {
-        vi.useFakeTimers({ now: START });
-        const card = mountCycling();
-        startReplayJustBeforeATick(card);
-
-        vi.advanceTimersByTime(59900); // lands exactly on the next tick, replay long finished
-        expect(card._dateNav.isReplaying).toBe(false);
-        expect(card._dateNav.currentDate.getTime()).toBe(Date.now());
-        expect(card._zoomLevel).not.toBe(1);
-        card.remove();
-      });
-    });
-  });
-
-  describe("manual interaction suspends the periodic zoom cycle", () => {
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    function mountCycling() {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      return card;
-    }
-
-    for (const action of [
-      "hour-back",
-      "hour-forward",
-      "day-back",
-      "day-forward",
-      "month-back",
-      "month-forward",
-    ]) {
-      it(`${action} suspends the cycle until Home`, () => {
-        vi.useFakeTimers();
-        const card = mountCycling();
-        clickButton(card, action);
-        vi.advanceTimersByTime(60000);
-        expect(card._zoomLevel).toBe(1);
-
-        clickButton(card, "today");
-        vi.advanceTimersByTime(60000);
-        expect(card._zoomLevel).toBe(2);
-        card.remove();
-      });
-    }
-
-    it("dragging the view suspends the cycle", () => {
-      vi.useFakeTimers();
-      const card = mountCycling();
-      dragView(card, 50, 0);
-
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(1);
-      card.remove();
-    });
-
-    it("replay does not suspend the cycle", () => {
-      vi.useFakeTimers();
-      const card = mountCycling();
-      clickButton(card, "replay");
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      card.remove();
-    });
-
-    it("opening the gallery does not suspend the cycle", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true, gallery: { mode: "both" } });
-      document.body.appendChild(card);
-      clickButton(card, "gallery");
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      card.remove();
-    });
-  });
-
-  describe("periodic_zoom_max configuration", () => {
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    it("wraps at configured max level instead of MAX_ZOOM", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true, periodic_zoom_max: 3 });
-      document.body.appendChild(card);
-      expect(card._zoomLevel).toBe(1);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(2);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(3);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(1);
-      card.remove();
-    });
-
-    it("defaults to MAX_ZOOM (4) when periodic_zoom_max is not set", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true });
-      document.body.appendChild(card);
-      vi.advanceTimersByTime(60000);
-      vi.advanceTimersByTime(60000);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(4);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(1);
-      card.remove();
-    });
-
-    it("defaults to MAX_ZOOM for invalid periodic_zoom_max values", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_max: "abc" });
-      expect(card._zoom.periodicZoomMax).toBe(4);
-      card.setConfig({ periodic_zoom_max: 2.5 });
-      expect(card._zoom.periodicZoomMax).toBe(4);
-      card.setConfig({ periodic_zoom_max: 1 });
-      expect(card._zoom.periodicZoomMax).toBe(4);
-      card.setConfig({ periodic_zoom_max: 5 });
-      expect(card._zoom.periodicZoomMax).toBe(4);
-    });
-
-    it("has no effect when periodic_zoom_change is false", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: false, periodic_zoom_max: 3 });
-      document.body.appendChild(card);
-      expect(card._zoomLevel).toBe(1);
-      vi.advanceTimersByTime(60000);
-      expect(card._zoomLevel).toBe(1);
       card.remove();
     });
   });
@@ -1554,111 +1080,6 @@ describe("SolarViewCard", () => {
       svg.dispatchEvent(
         new PointerEvent("pointerup", { clientX: 250, clientY: 200, pointerId: 1 })
       );
-      card.remove();
-    });
-  });
-
-  describe("auto-update timer", () => {
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    it("starts a timer on connectedCallback", () => {
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      expect(card._autoUpdateTimer).toBeNull();
-      document.body.appendChild(card);
-      expect(card._autoUpdateTimer).not.toBeNull();
-      card.remove();
-    });
-
-    it("clears timer on disconnectedCallback", () => {
-      const card = createAndMount();
-      expect(card._autoUpdateTimer).not.toBeNull();
-      card.remove();
-      expect(card._autoUpdateTimer).toBeNull();
-    });
-
-    it("clears existing timer before creating new one on reconnect", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      document.body.appendChild(card);
-      const firstTimer = card._autoUpdateTimer;
-      card.remove();
-      document.body.appendChild(card);
-      const secondTimer = card._autoUpdateTimer;
-      expect(secondTimer).not.toBeNull();
-      expect(secondTimer).not.toBe(firstTimer);
-      card.remove();
-    });
-
-    it("re-renders after 60s when showing today", () => {
-      vi.useFakeTimers({ now: new Date("2026-02-15T10:00:00") });
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card._dateNav.currentDate = new Date("2026-02-15T10:00:00");
-      document.body.appendChild(card);
-      vi.advanceTimersByTime(60000);
-      // Date should have been updated to "now" (still Feb 15)
-      expect(formatDate(card._dateNav.currentDate)).toContain("26-02-15");
-      card.remove();
-    });
-
-    it("syncs to current time on visibilitychange when in live mode", () => {
-      vi.useFakeTimers({ now: new Date("2026-02-15T10:00:00") });
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      document.body.appendChild(card);
-      // Simulate time passing while tab was hidden (timer throttled)
-      vi.setSystemTime(new Date("2026-02-15T10:45:00"));
-      // Card still shows stale time — timer never fired
-      expect(formatDate(card._dateNav.currentDate)).toContain("10:00");
-      // Tab becomes visible
-      Object.defineProperty(document, "hidden", { value: false, configurable: true });
-      document.dispatchEvent(new Event("visibilitychange"));
-      // Card should now show current time
-      expect(formatDate(card._dateNav.currentDate)).toContain("10:45");
-      card.remove();
-    });
-
-    it("does not sync on visibilitychange when user has navigated away", () => {
-      vi.useFakeTimers({ now: new Date("2026-02-15T10:00:00") });
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      document.body.appendChild(card);
-      card._navigate(-45 * 86400000); // user went to a past date
-      vi.setSystemTime(new Date("2026-02-15T10:45:00"));
-      Object.defineProperty(document, "hidden", { value: false, configurable: true });
-      document.dispatchEvent(new Event("visibilitychange"));
-      // Should still show the navigated-to date
-      expect(formatDate(card._dateNav.currentDate)).not.toContain("26-02-15");
-      card.remove();
-    });
-
-    it("does not sync on visibilitychange while the tab is still hidden", () => {
-      vi.useFakeTimers({ now: new Date("2026-02-15T10:00:00") });
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      document.body.appendChild(card);
-      vi.setSystemTime(new Date("2026-02-15T10:45:00"));
-      Object.defineProperty(document, "hidden", { value: true, configurable: true });
-      document.dispatchEvent(new Event("visibilitychange"));
-      expect(formatDate(card._dateNav.currentDate)).toContain("10:00");
-      Object.defineProperty(document, "hidden", { value: false, configurable: true });
-      card.remove();
-    });
-
-    it("removes visibilitychange listener on disconnectedCallback", () => {
-      const card = createAndMount();
-      expect(card._onVisibilityChange).not.toBeNull();
-      card.remove();
-      expect(card._onVisibilityChange).toBeNull();
-    });
-
-    it("does not re-render when user has navigated to a different date", () => {
-      vi.useFakeTimers({ now: new Date("2026-02-15T10:00:00") });
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      document.body.appendChild(card);
-      // Navigate backwards 45 days via the proper navigation path, which sets _isLiveMode=false
-      card._navigate(-45 * 86400000);
-      vi.advanceTimersByTime(60000);
-      // Should still show the navigated-to date, not auto-advance to today
-      expect(formatDate(card._dateNav.currentDate)).not.toContain("26-02-15");
       card.remove();
     });
   });
@@ -1841,24 +1262,6 @@ describe("SolarViewCard", () => {
       card.setConfig({ default_zoom: 3 });
       expect(card._zoomLevel).toBe(3);
       expect(parseViewBox(card).width).toBe(480);
-      card.remove();
-    });
-  });
-
-  describe("animated zoom with periodic auto-cycle", () => {
-    afterEach(() => {
-      vi.useRealTimers();
-    });
-
-    it("auto-cycle updates zoom level display with animation enabled", () => {
-      vi.useFakeTimers();
-      const card = document.createElement("ha-planetary-solar-system-card-test");
-      card.setConfig({ periodic_zoom_change: true, zoom_animate: true });
-      document.body.appendChild(card);
-      expect(card.shadowRoot.querySelector(".zoom-level").textContent).toBe("1");
-      vi.advanceTimersByTime(60000);
-      // Zoom level display should update immediately to target
-      expect(card.shadowRoot.querySelector(".zoom-level").textContent).toBe("2");
       card.remove();
     });
   });
